Add table components
Fix the key to `school.name` as that is unique and preferred over an array index. We're not manipulating the DOM here to really worry but this is a better practice.
Add index to help with setting even row colors and our "prop hell" of passing athlete_gpa through to where it's needed. We'll clean this up with Pinia later.

This is a 1:1 representation of what we had before and appears accurate. This is where I'll start to add computed properties, format the GPA to 2 decimals, etc.