Normalize GPA and colorize
Format GPA in the table to 2 decimal places.
Tailwind usually has a problem purging dynamic classes like this but since it's in the root and all values are present, I don't personally mind this shortcut. The alternative of having all the logic in the template would be verbose and cumbersome. Not the biggest fan of the logic but I'm running out of time.