Last minute cleanup
Remove Tailwind's even usage in favor of explicitly setting it. Now the sticky header has a background color and doesn't look so janky being transparent.
Use flex-shrink-0 to make sure the collapsed header sections do not shrink. One large conference value was pushing just one item inward, looking very out of place.
Fix formatting on the AcademicFitReport test as I forgot it apparently.
Add a test covering that the name can be edited.
Add the Pinia section to the readme to show I'm skipping it for time. I planned on integrating it last.